# Add cheap ST_Intersection method for certain cases